script/process_fig_files.requirements.txt
gradio_client
pillow
